+++ tests/dummy/config/environment.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+
 module.exports = function (environment) {
   var ENV = {
     contentSecurityPolicy: {
@@ -15,11 +17,11 @@ module.exports = function (environment) {
       FEATURES: {
         // Here you can enable experimental features on an ember canary build
         // e.g. 'with-controller': true
+      },
+      EXTEND_PROTOTYPES: {
+        // Prevent Ember Data from overriding Date.parse.
+        Date: false
       }
-      // TODO Do we want to disable prototype extension?
-      // EXTEND_PROTOTYPES: {
-      //   Date: false
-      // }
     },
 
     APP: {
@@ -39,7 +41,6 @@ module.exports = function (environment) {
 
   if (environment === 'test') {
     // Testem prefers this...
-    ENV.rootURL = '/'
     ENV.locationType = 'none'
 
     // keep test console output quieter