+++ tests/integration/components/frost-object-browser-test.js
@@ -1,18 +1,17 @@
-/**
- * Integration test for the frost-object-browser component
- */
-
 import {expect} from 'chai'
+import {$hook, initialize as initializeHook} from 'ember-hook'
 import {registerMockComponent, unregisterMockComponent} from 'ember-test-utils/test-support/mock-component'
 import {integration} from 'ember-test-utils/test-support/setup-component-test'
 import hbs from 'htmlbars-inline-precompile'
 import {afterEach, beforeEach, describe, it} from 'mocha'
+import sinon from 'sinon'
 
 const test = integration('frost-object-browser')
 describe(test.label, function () {
   test.setup()
 
   beforeEach(function () {
+    initializeHook()
     registerMockComponent(this, 'mock-content')
     registerMockComponent(this, 'mock-controls')
     registerMockComponent(this, 'mock-filters')
@@ -55,6 +54,14 @@ describe(test.label, function () {
     it('should render the refine by label in the facets', function () {
       expect(this.$('.frost-object-browser-facets-header')).to.have.text('Refine by')
     })
+
+    it('should display facets-expanded-view without isFilterHiddenOnLoad', function () {
+      expect($hook('myObjectBrowser-facets-expanded-view').is(':visible')).to.eql(true)
+    })
+
+    it('should hide facets-collapsed-view without isFilterHiddenOnLoad', function () {
+      expect($hook('myObjectBrowser-facets-collapsed-view').is(':visible')).to.eql(false)
+    })
   })
 
   describe('after render, when given a new refineByLabel', function () {
@@ -87,6 +94,140 @@ describe(test.label, function () {
 
     it('should render the refine by label in the facets', function () {
       expect(this.$('.frost-object-browser-facets-header')).to.have.text('Filtrar por')
+    })
+  })
+
+  describe('when rendering with isFilterHiddenOnLoad=false', function () {
+    beforeEach(function () {
+      this.set('isFilterHiddenOnLoad', false)
+
+      this.render(hbs`
+        {{frost-object-browser
+          hook=myHook
+          hookQualifiers=(hash)
+          content=(component 'mock-content' class='mock-content')
+          controls=(component 'mock-controls' class='mock-controls')
+          filters=(component 'mock-filters' class='mock-filters')
+          isFilterHiddenOnLoad=isFilterHiddenOnLoad
+        }}
+      `)
+    })
+
+    it('should display facets-expanded-view', function () {
+      expect($hook('myObjectBrowser-facets-expanded-view').is(':visible')).to.eql(true)
+    })
+
+    it('should hide facets-collapsed-view', function () {
+      expect($hook('myObjectBrowser-facets-collapsed-view').is(':visible')).to.eql(false)
+    })
+
+    describe('when click filter collapse icon', function () {
+      beforeEach(function () {
+        return $hook('myObjectBrowser-facets-collapse-control').click()
+      })
+
+      it('should hide facets-expanded-view', function () {
+        expect($hook('myObjectBrowser-facets-expanded-view').is(':visible')).to.eql(false)
+      })
+
+      it('should display facets-collapsed-view', function () {
+        expect($hook('myObjectBrowser-facets-collapsed-view').is(':visible')).to.eql(true)
+      })
+    })
+  })
+
+  describe('when rendering with isFilterHiddenOnLoad=true', function () {
+    beforeEach(function () {
+      this.set('isFilterHiddenOnLoad', true)
+
+      return this.render(hbs`
+        {{frost-object-browser
+          hook=myHook
+          hookQualifiers=(hash)
+          content=(component 'mock-content' class='mock-content')
+          controls=(component 'mock-controls' class='mock-controls')
+          filters=(component 'mock-filters' class='mock-filters')
+          isFilterHiddenOnLoad=isFilterHiddenOnLoad
+        }}
+      `)
+    })
+
+    it('should hide facets-expanded-view', function () {
+      expect($hook('myObjectBrowser-facets-expanded-view').is(':visible')).to.eql(false)
+    })
+
+    it('should display facets-collapsed-view', function () {
+      expect($hook('myObjectBrowser-facets-collapsed-view').is(':visible')).to.eql(true)
+    })
+
+    describe('when click filter expand icon', function () {
+      beforeEach(function () {
+        return $hook('myObjectBrowser-facets-expand-control').click()
+      })
+
+      it('should display facets-expanded-view', function () {
+        expect($hook('myObjectBrowser-facets-expanded-view').is(':visible')).to.eql(true)
+      })
+
+      it('should hide facets-collapsed-view', function () {
+        expect($hook('myObjectBrowser-facets-collapsed-view').is(':visible')).to.eql(false)
+      })
+    })
+  })
+
+  describe('when user subscribe to onFilterHide event', function () {
+    let hideFilter
+    beforeEach(function () {
+      hideFilter = sinon.spy()
+
+      this.set('isFilterHiddenOnLoad', false)
+      this.set('hideFilter', hideFilter)
+
+      return this.render(hbs`
+        {{frost-object-browser
+          hook=myHook
+          hookQualifiers=(hash)
+          content=(component 'mock-content' class='mock-content')
+          controls=(component 'mock-controls' class='mock-controls')
+          filters=(component 'mock-filters' class='mock-filters')
+          isFilterHiddenOnLoad=isFilterHiddenOnLoad
+          onFilterHide=(action hideFilter)
+        }}
+      `)
+    })
+
+    it('should call the handler hideFilter', function () {
+      $hook('myObjectBrowser-facets-collapse-control').click()
+
+      expect(hideFilter).to.have.callCount(1)
+    })
+  })
+
+  describe('when user subscribe to onFilterDisplay event', function () {
+    let displayFilter
+    beforeEach(function () {
+      displayFilter = sinon.spy()
+
+      this.set('isFilterHiddenOnLoad', false)
+      this.set('displayFilter', displayFilter)
+
+      return this.render(hbs`
+        {{frost-object-browser
+          hook=myHook
+          hookQualifiers=(hash)
+          content=(component 'mock-content' class='mock-content')
+          controls=(component 'mock-controls' class='mock-controls')
+          filters=(component 'mock-filters' class='mock-filters')
+          isFilterHiddenOnLoad=isFilterHiddenOnLoad
+          onFilterDisplay=(action displayFilter)
+        }}
+      `)
+    })
+
+    it('should call the handler displayFilter', function () {
+      $hook('myObjectBrowser-facets-expand-control').click()
+
+      expect(displayFilter).to.have.callCount(1)
     })
   })
 })